+++ src/kmeans_cuda_shared.h
@@ -1,13 +1,13 @@
 #include "argparse.h"
 #include <cuda_runtime.h>
 
-void kmeans_cuda_shared(double *dataset, double * centroids, options_t &args);
-double * cuda_shared_copy(double * original, options_t args);
+void kmeans_cuda_shared(float *dataset, float * centroids, options_t &args);
+float * cuda_shared_copy(float * original, options_t args);
 
-bool cuda_shared_converged(double * new_centroids, double* old_centroids, options_t &args, double * duration);
-__global__ void d_cuda_shared_convergence_helper(double * new_c, double * old_c, bool * convergence, double threshold, int dimensions);
-int * cuda_shared_find_nearest_centroids(double * dataset, double * centroids, options_t &args, double * duration);
-__global__ void d_cuda_shared_find_nearest_centroids(double * dataset, double * centroids, int * labels, int dims, double max);
-double * cuda_shared_average_labeled_centroids(double * dataset, int * labels, options_t &args, double * duration);
-__global__ void d_cuda_shared_average_labeled_centroids(double * d_dataset, int * d_labels, double * centroids, int number_of_values);
+bool cuda_shared_converged(float * new_centroids, float* old_centroids, options_t &args, float * duration);
+__global__ void d_cuda_shared_convergence_helper(float * new_c, float * old_c, bool * convergence, float threshold, int dimensions);
+int * cuda_shared_find_nearest_centroids(float * dataset, float * centroids, options_t &args, float * duration);
+__global__ void d_cuda_shared_find_nearest_centroids(float * dataset, float * centroids, int * labels, int dims, float max);
+float * cuda_shared_average_labeled_centroids(float * dataset, int * labels, options_t &args, float * duration);
+__global__ void d_cuda_shared_average_labeled_centroids(float * d_dataset, int * d_labels, float * centroids, int number_of_values);
